# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AeroNode API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{id}/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks in a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task in a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (optimistic concurrency via version)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "409":
          description: Version conflict — someone else already updated this task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionConflict"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    TaskStatus:
      type: string
      enum: [todo, in_progress, done]
    TaskPriority:
      type: string
      enum: [low, medium, high, urgent]
...
    Task:
      type: object
      properties:
        id: { type: integer }
        workspaceId: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        dueDate: { type: ["string", "null"], format: date-time }
        order: { type: integer }
        assigneeId: { type: ["integer", "null"] }
        assigneeName: { type: ["string", "null"] }
        createdBy: { type: integer }
        lockedBy: { type: ["integer", "null"] }
        lockedByName: { type: ["string", "null"] }
        version: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        [id, workspaceId, title, description, status, priority, dueDate, order, assigneeId, assigneeName,
         createdBy, lockedBy, lockedByName, version, createdAt, updatedAt]
    TaskInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        dueDate: { type: ["string", "null"], format: date-time }
        assigneeId: { type: ["integer", "null"] }
      required: [title]
    TaskUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        dueDate: { type: ["string", "null"], format: date-time }
        order: { type: integer }
        assigneeId: { type: ["integer", "null"] }
        version: { type: integer }
      required: [version]
    VersionConflict:
      type: object
      properties:
        error: { type: string }
        task:
          $ref: "#/components/schemas/Task"
      required: [error, task]
...